refactor(planning): test files realocation
